edit desgin of transfer page

+++ nkba/app/Http/Controllers/Admin/ADoctorsController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use DB;
+use Session;
 use App\Doctor;
 use App\Http\Requests;
 use App\Http\Controllers\Controller;
@@ -11,6 +12,7 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Input;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\Redirect;
+use Illuminate\Support\Facades\Auth;
 
 class ADoctorsController extends Controller
 {
@@ -21,9 +23,14 @@ class ADoctorsController extends Controller
      */
     public function index()
     {
-        $doctors = Doctor::all();
-        $doctors->toarray();
-        return view('admin.doctors.index', compact('doctors'));
+        if (!Session::has('id')) {
+            return redirect("/admin/login");
+        }else{
+            $doctors = Doctor::all();
+            $doctors->toarray();
+            return view('admin.doctors.index', compact('doctors'));
+        }
+
     }
 
     /**
@@ -33,7 +40,11 @@ class ADoctorsController extends Controller
      */
     public function create()
     {
-        return view('admin.doctors.new');
+        if (!Session::has('id')) {
+            return redirect("/admin/login");
+        }else{
+            return view('admin.doctors.new');
+        }
     }
 
     /**
@@ -44,26 +55,40 @@ class ADoctorsController extends Controller
      */
     public function store(Request $request)
     {
-        $input = Input::all();
+        if (!Session::has('id')) {
+            return redirect("/admin/login");
+        }else{
+            $input = Input::all();
         // $validation = Validator::make($input, Doctor::$rules);
         // if ($validation->passes())
         // {
             //add doctor in doctors table
-        $doctor = new Doctor;
-        $doctor->name = $input['name'];
-        $doctor->degree = $input['degree'];
-        $doctor->specialization = $input['specialization'];
-        $doctor->address = $input['address'];
-        $doctor->phone = $input['phone'];
-        $doctor->governorate = $input['governorate'];
-        $doctor->area = $input['area'];
-        $doctor->price = $input['price'];
-        $doctor->mobile = $input['mobile'];
-        $doctor->discription = $input['discription'];
-        $doctor->path = $input['path'];
-        $doctor->save();
 
-        return Redirect::route('admin-doctor.index');
+            $destinatonPath = '';
+            $photoname = '';
+            if(Input::file('path')){
+                $photo = Input::file('path');
+                $destinationPath = '/assets/images/admin/doctors/';
+                $extension = Input::file('path')->getClientOriginalExtension();
+                $photoname = mt_rand(1, 100000).$photo->getClientOriginalName();
+                $photo->move($destinationPath,$photoname);
+            }
+            $doctor = new Doctor;
+            $doctor->name = $input['name'];
+            $doctor->degree = $input['degree'];
+            $doctor->specialization = $input['specialization'];
+            $doctor->address = $input['address'];
+            $doctor->phone = $input['phone'];
+            $doctor->governorate = $input['governorate'];
+            $doctor->area = $input['area'];
+            $doctor->price = $input['price'];
+            $doctor->mobile = $input['mobile'];
+            $doctor->discription = $input['discription'];
+            $doctor->path = $destinationPath.$photoname;
+            $doctor->save();
+
+            return Redirect::route('admin-doctor.index');
+        }
         // } 
 
         // return Redirect::route('admin-doctor.create')
@@ -81,8 +106,12 @@ class ADoctorsController extends Controller
     public function show($id)
     {
         // get the doctor
-        $doctor = Doctor::find($id);
-        return View('admin.doctors.show',compact('doctor'));
+        if (!Session::has('id')) {
+            return redirect("/admin/login");
+        }else{
+            $doctor = Doctor::find($id);
+            return View('admin.doctors.show',compact('doctor'));
+        }
     }
 
     /**
@@ -93,8 +122,12 @@ class ADoctorsController extends Controller
      */
     public function edit($id)
     {
-        $doctor = Doctor::find($id);
-        return View('admin.doctors.edit',compact('doctor'));
+        if (!Session::has('id')) {
+            return redirect("/admin/login");
+        }else{
+            $doctor = Doctor::find($id);
+            return View('admin.doctors.edit',compact('doctor'));
+        }
     }
 
     /**
@@ -106,14 +139,18 @@ class ADoctorsController extends Controller
      */
     public function update(Request $request, $id)
     {
-        $doctor = Doctor::findOrFail($id);
+        if (!Session::has('id')) {
+            return redirect("/admin/login");
+        }else{
+            $doctor = Doctor::findOrFail($id);
        // $validation=$this->validate($request, [
        //  'name' => 'required',
        //  'email' => 'required'
        //  ]);
-        $input = $request->all();
-        $doctor->fill($input)->save();
-        return Redirect::route('admin-doctor.show',$doctor->id);
+            $input = $request->all();
+            $doctor->fill($input)->save();
+            return Redirect::route('admin-doctor.show',$doctor->id);
+        }
        // ->withInput()
        // ->withErrors($validation)
        // ->with('message', 'There were validation errors.');
@@ -127,8 +164,12 @@ class ADoctorsController extends Controller
      */
     public function destroy($id)
     {
-        $doctor = Doctor::findOrFail($id);
-        $doctor->delete();
-        return Redirect::route('admin-doctor.index');
+        if (!Session::has('id')) {
+            return redirect("/admin/login");
+        }else{
+            $doctor = Doctor::findOrFail($id);
+            $doctor->delete();
+            return Redirect::route('admin-doctor.index');
+        }
     }
 }
